✨ 새 기능 :  코인 업데이트 시 매수 매도 주문 체결

+++ backend/src/main/java/com/kmbbj/backend/feature/exchange/repository/cassandra/sell/SellOrderRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 
 @Repository
 public interface SellOrderRepository extends CassandraRepository<SellOrder, OrderPrimaryKey> {
-    List<SellOrder> findAllByIdCoinIdAndIdPriceLessThanEqual(Long coinId, BigDecimal price);
+    List<SellOrder> findAllByIdCoinIdAndIdPriceLessThanEqual(Long coinId, Long price);
     void deleteByIdCoinIdAndTransactionId(Long coinId, Long transactionId);
 }
