+++ backend/src/main/java/com/kmbbj/backend/feature/exchange/service/execution/matching/ExecutionAllMatchingOrder.java
@@ -1,7 +1,5 @@
 package com.kmbbj.backend.feature.exchange.service.execution.matching;
 
-import java.math.BigDecimal;
-
 public interface ExecutionAllMatchingOrder {
-    void matchOrders(Long coinId, BigDecimal price);
+    void matchOrders(Long coinId, Long price);
 }
